+++ Examples/Example08.ps1
@@ -32,6 +32,7 @@ $RootPath = $PSScriptRoot
 <#
     The $Images variable is a hashtable containing the names of image files used in the diagram.
     The image files must be located in the directory specified by $IconPath.
+    ** Image sizes should be around 100x100, 150x150 pixels for optimal display. **
 #>
 
 $script:Images = @{
@@ -71,7 +72,7 @@ $example8 = & {
         https://psgraph.readthedocs.io/en/latest/Command-SubGraph/
     #>
 
-    SubGraph 3tier -Attributes @{Label = '3Tier Concept'; fontsize = 22; penwidth = 1.5; labelloc = 't'; style = "dashed,rounded"; color = "darkgray" } {
+    SubGraph 3tier -Attributes @{Label = '3 Tier Concept'; fontsize = 22; penwidth = 1.5; labelloc = 't'; style = "dashed,rounded"; color = "darkgray" } {
 
         $WebServerFarm = @(
             @{